Update on Calendar 9
Okay I figured out how to get the modal pop into the frame when the button is click the next step is properly implementing it into the code for saving events and clicking on events from the calendar as well.

I also decided that to change the calendar format so it also features the month which will hopefully help out a lot of users as well!