Update on Calendar 17
Final update, I added in the comments for all the files and deleted any extra code that is no longer needed within in those files. I also deleted the old files that are no longer relevant. IMPORTANT NOTICE: just because the files are called SQlite there are no databases relating back to them, I hope this makes sense. This should also be the final update for now!